fly: behaviour: accept arbitrary length tokens, take 3
I'm surprised how difficult it's proven to be to just read input from
the terminal without hitting buffer limits. The unit tests on our
Windows worker highlighted some undesirable behaviour with the
peterh/liner library. It didn't seem to successfully detect the number
of columns (at least on Windows), which caused:

a) The password prompt to fail
b) The regular prompt, well, not printing the prompt
c) Not using raw mode, which defeats the whole purpose of this fix!

I'm not sure exactly why the columns were detected to be 0, and why the
library even cares about the number of columns for something as simple
as our use case. peterh/liner does a lot of cool stuff, but we were
using practically none of it - so why not go with a simpler solution
that we control!

In place of peterh/liner, just use a raw mode terminal + our own line
reading logic. Unfortunately, since we're using raw mode, new lines
don't also come with carriage returns, so I needed to manually add them
throughout the print statements.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>