increase timeout for system claim startup error
I've noticed this test flaking for a while now, and it's integrating across a
wide boundary -- I think it's OK to simply make it a bit more permissive.

On the other hand, it's a very small piece of validation logic that is being
tested, which would be a natural candidate for a unit test. We may have a bit
of a separation problem here.

On the other other hand, we're going to revisit configuration soon so it might
make sense to pay down this pre-existing tech debt later when we're doing the
right thing anyway.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>