web: behaviour: offset sidebar tooltips
concourse/concourse#5435

With overflowing pipeline names, hovering over the star icon brings up
the tooltip which then covers the star icon. Since the z-index of the
tooltip is greater than the sidebar pipeline, we're no longer hovering
the pipeline, so the tooltip goes away. This cycle repeats, causing the
tooltip to flicker really quickly (and block the star icon).

One solution would be to move the DOM ID of the pipeline name to the
entire row (which would include the star icon) so that the tooltip
appears to the right of the pipeline wrapper (and therefore to the right
of the star icon). The problem with this is that the DOM ID needs to be
on the text node itself, since otherwise text overflowing doesn't work,
so the tooltip is never displayed and the text is not ellipsized (see #5855).

The hacky solution we came up with was to manually offset the tooltip by
a fixed amount. This will likely be pretty brittle to style changes, but
it's the best we could come up with.

Signed-off-by: Zoe Tian <ztian@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>