fly: always warn on set-pipeline if pipeline is paused
Users may have been confused when executing `set-pipeline` repeatedly
and still seeing no builds running. The note about the pipeline being
paused was printed only when the pipeline is initially set, not on
subsequent updates to the config.

The new behavior prints the message on every successful set-pipeline
invocation, if and only if the pipeline is paused. That will certainly
be the case for freshly-created pipelines, and may also be true for
other pipelines later in their lives.

Signed-off-by: Alexander Gurney <alexander_gurney@cable.comcast.com>