web: behaviour: render across step in build plan
The across step is initialized and rendered differently from any
existing step, in that it contains other steps but is itself a step that
needs to receive events and be rendered as an expandable step header.
Previously, no multi-step (e.g. `in_parallel`, `do`, `retry` etc.) had
its own step header.

Specifically, the across step has its own header that indicates the
variable name we're iterating over. Clicking into the across step, you
see both the logs that indicate the step is experimental, and a list of
tabs (like the `retry` step) that display all of the values that the
variable took on. Each tab has a coloured line above indicating the
status of the build for the corresponding value.

This structure is much like the build plan that gets returned - an
`across` step with multiple variables gets rendered as nested across
steps.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>