atc: behaviour: track users on login
...rather than on every request.

concourse/concourse#5851

Now that we're back to creating our own access tokens, we can simplify
the user tracking behaviour. With the auth changes to 6.1, in order to
track user activity, we updated the "users.last_login" field in the DB
on every request. Since that was found to be pretty slow, we added an
optimization where we keep track of user updates in memory, and every 15
seconds (or if the number of users exceeds some limit), dump that to the
database.

In an effort to remove complexity from auth, just track the "last login"
time when users actually login

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>