atc/steps: Don't allow modifier only steps
A step that consists of only modifiers (e.g. attempts, , on_success,
etc.) is an invalid step. This commit ensures that an error is returned
if no core step types are specified.

From what we can tell, StepWrapper == step modifier.

We updated the comments to hopefully make it easier to understand how
step parsing works.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: James Thomson <jthomson@pivotal.io>