worker: refactor: make clearer use of garden terminology
Garden provides two things - a server with API endpoints
and an interface for a backend that has to implemented in
order to use that server meaningfully. The component imple-
menting the backend is a runtime. This commit tries to
make clearer use of these terms where appropriate and intro-
duces comments to reduce ambiguity.

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>
Signed-off-by: Muntasir Chowdhury <muntasir.mzc@gmail.com>