worker: rename deny-network to restricted-network for containerd
restricted-network works semantically better as it refers to
a network that is restricted. deny-network suggests an action
whose value could either be "true" or "false". The deny-network
terminology is from Guardian and was initially used to avoid a
new CONCOURSE_GARDEN flag. Since we are introducing a separate
flag for this feature now anyway it is a good time to rename
it to CONCOURSE_CONTAINERD_RESTRICTED_NETWORK.

Signed-off-by: Muntasir Chowdhury <mchowdhury@pivotal.io>