web: behaviour: revamp across step UI
Rather than displaying across steps over multiple vars as nested across
steps, render a flat list of headers that contain the key-value pairs
for all the vars set for a particular iteration. For instance, if
running `across` 2 vars (`var1` and `var2`), where each var takes on 2
different values, you'd see 4 headers nested under the `across `header`
that are labelled with "var1: a1   var2: b1", etc.

This commit introduces a pseudo "sub step" header, that is a header that
isn't a step itself, but wraps a step. Since it's not a step, we need to
determine the state by traversing its substep.

I also realized there was a bug in the `mostSevereStepState` function -
if you had a retry step which succeeded without going through all of the
attempts or a hook that did not get invoked because the step succeeded,
the "most severe status' would be pending rather than succeeded (since
pending has priority over succeeded, since for a step tree to have
succeeded, everything under it must have succeeded). To address this, I
modified to `fold` function to not traverse step trees that will never
be executed.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>