web: behaviour: fix scroll to highlight for nested step
offsetHeight just considers the offsetParent (the closest parent element
with position: relative or something along those lines). When we have a
step nested in another step, the offsetParent won't be the same, so just
using offsetHeight is not the height of the line within the entire build
body. To solve this, you have to traverse the offsetParents and sum up
their offset heights.

To avoid the highlighted line appearing behind the step headers (as much
as possible), increase the padding significantly. This way, if we have 3
stacked step headers (as we do with the across step), the highlighted
line isn't hidden behind.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>