web: behaviour: delay sticky header syncing
There seems to be a race with rendering the newly expanded step headers
and attempting to sync their heights - wasn't super common, but happened
to be a few times. I assumed the setTimeout(0) would fix this, but
apparently not consistently. Increase the timeout to 50ms, which seems
to work more consistently.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>