fly: structure: facilitate constructing flyrc in tests
...by exposing the RC struct

Constructing flyrc's in tests means we don't have to use fixtures. This
is useful, since it's easier to programmatically generate the access
tokens if we do in fact need to look at the "opaque" access tokens to
infer the expiry time (which we will need to do to maintain parity with
the current featureset of `fly targets`).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>