fly: behaviour: targets command works with new token
It's unfortunate, but in order to maintain parity with earlier versions
of `fly targets` (that report on the expiry of the access token), we
can't treat the access token as opaque. This breaks the `Bearer` token
contract for clients. However, `fly` is sort of a "special" client
anyway, in that it's bundled with Concourse itself, and is versioned
alongside it.

In #5897, we encoded the expiry into the access token, making this
feature easy to implement.

This commit also adds a helper for generating a `.flyrc` that can be
useful in other `fly` integration tests as well.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>