atc/metric: structure: add events test helper
Because `Invocations()` involves the acquisition of many locks (one for each
method), multiple parallel calls to `Invocations()` can result in the following:
one actor acquires a lock for one method and another actor acquires the lock for
a different method, and then they deadlock waiting for each other to unlock.

With this change, that deadlock can be effectively avoided, since internally all
contention is for the `emitMutex`.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>