db: add migration for deleted_pipelines table
concourse/concourse#5287

This migration also removes the trigger to delete the
pipeline_build_events_x table when a pipeline is deleted.

Note that we want to keep the trigger for deleting teams because one-off
builds don't get deleted in the build log collector, so there is no risk
of deadlock there.

Signed-off-by: Zoe Tian <ztian@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>