web: always show all pipelines header on dashboard
For consistency with the sidebar, always show the all pipelines header.
This resulted in a bunch of tests failing since the extra offset from
the header caused pipeline cards not to be rendered. As we've done
before, replace each `Common.init` with `whenOnDashboard` which sets up
an initial screen size.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>