fly: get-team --json actually produces JSON
This is documented, implemented-ish, and tested-ish, but didn't work:
doing `fly -t tgt get-team --team-name foo --json` would always produce
the valid but boring response `{}`. This is because the code uses
the go-concourse `Team` object, which has no data members and so does
very little of interest when serialized.

There was originally a test for the YAML-formatted output, but that
did not run for the reasons described in
https://github.com/concourse/concourse/commit/a26d9e09, and there were
no tests for JSON. That commit also changed the test to look at the
tabular format. It appears tests of the JSON path were accidentally
skipped over in the refactoring.

We now synthesize a JSON object with team and auth contents, which
almost matches the format used elsewhere - what's missing is the ID,
which isn't currently plumbed through the go-concourse client logic.
Also, the new code has a unit test.

Signed-off-by: Alexander Gurney <alexander_gurney@cable.comcast.com>