atc: Allow OPA integration to return reasons
Allow OPA documents to include an `allowed` and a `reasons` field that
gives a bit more feedback to the user about why something failed. This
allows OPA operators to write rules that are self explanatory as to why
things are getting denied.

Fly does not appear to print the full error message currently but this
works for the UI.

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>