Enable garden DNS proxy in docker-compose
The `CONCOURSE_GARDEN_DNS_PROXY_ENABLE` flag was taken out of the dev
image in https://github.com/concourse/ci/commit/db8c43e565777d767b6f4940573b6df2329d5865
In order for local docker-compose DNS to work with latest concourse/dev
image we need to renable it. Since Guardian is still the default in
docker-compose.yml it makes sense to use it there and unset it in the
containerd override.

Signed-off-by: Andy Paine <andy.paine@engineerbetter.com>