atc: structure: beginnings of dbtest package
this package is intended to clean up all the messy setup code throughout
the db tests.

its development is spurred on by the fact that .SetResourceConfig is
used throughout the tests, and is no longer used in production, and
updating all of the tests is going to take a lot of time.

this package hopes to provide a higher level interface for setting up
realistic test scenarios, so that we may refactor the underlying
modeling in only one place - the dbtest package.

initially, only a few tests here have been updated to show interesting
scenarios. the following commits will keep pulling on the
.SetResourceConfig thread until it's all unraveled.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>