atc: behaviour: simplify pagination
concourse/concourse#5977

Pagination was previously controlled primarily using `Since` and
`Until` which are exclusive range and view the data from newest to oldest

e.g. Until 4 means from the most recent ID to build ID 5

This was pretty confusing, and led to some bugs (like #5977).

This commit changes the language and view of the data. Pages are now
defined as an inclusive range `[From, To]` (side note: we previously had
`From` and `To`, but it was only used for resource versions and was
still newest to oldest, i.e. was `[To, From]`). Rather than speaking
about the next page and previous page, we found it easier to think about
when using "Newer" and "Older" page (since the next page is actually
older data). Note that we still refer to the pages as `next` and `prev`
in the API since those are standard refs.

Test changes will be included in a subsequent commit.

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>