atc: behaviour: diff default and zero for page
concourse/concourse#5977

If the FirstLoggedBuildID for a build was 0, we had no way to
differentiate between fetching `From 0` and not specifying a `From`
condition since we were using the zero-value to check if a field was
unset. Since 0 is a valid `From` condition, make `From` and `To`
pointers - this way, `nil` means unset, and `0` means 0.

Signed-off-by: Bohan Chen <bochen@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>