remove worker rebalancing test
here's the thing -- worker rebalancing has a misleading name. the worker doesn't
actually do anything to ensure that it connects to each TSA equally-often. in
the case of k8s, this only happens because the NodePort service type includes
a load-balancing feature. All the worker actually does is reconnect to its
gateway. Accordingly, there will always be a small (1/2^24 ish) chance that this
test will fail.

So let's not have our test suite assert on a nonexistent feature. We assume
that the reregistering logic is adequately tested in the worker modules. So this
test starts to look like a waste of compute time on CI since it is destined to
be flaky and tests at too wide of a boundary.

Signed-off-by: Jamie Klassen <jklassen@vmware.com>
Co-authored-by: Zoe Tian <twanru@vmware.com>