Update k8s-topgun for Helm v3
Behaviour around namespaces changed in Helm 3. Every helm command is now
scoped to a namespace just like kubectl. We now need to specify the
namespace in each helm command.

as of helm 3.2.0, we can add `--create-namespace` to the `helm upgrade`
command whic will allow helm to create the namespace if it doesn't
exist, following the same behaviour that was in helm2 versions.

Tiller is no longer required either, the `init` command was removed
because of this.

Signed-off-by: Vikram Yadav <vyadav@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Bishoy Youssef <byoussef@pivotal.io>